// client/src/model/Group/CreateGroup.ts
export type CreateGroupStudentsType = {
  course: number;
  number: string;
  educationType: string;
  speciality: string;
  receiptDate: Date;
  issueDate: Date;
  newStudent: string;
  students: string[];
};

export type CreateGroupType = {
  course: number;
  number: string;
  educationType: string;
  speciality: string;
  receiptDate: Date;
  issueDate: Date;
  newStudent: string;
};